% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_args.R
\docType{data}
\name{json_args}
\alias{json_args}
\title{json_args: test dataset for run_optimisation()}
\format{
A json list with all parameters used by \link{run_optimisation},
which are  passed to \link[kwb.heatsine]{run_optimisation}
}
\usage{
data(json_args)
}
\description{
test dataset generated with R script (data-raw/DATASET.R)
}
\examples{
json_args

}
\keyword{datasets}
