% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_optimisation.R
\name{run_optimisation}
\alias{run_optimisation}
\title{JSON wrapper function for kwb.heatsine::run_optimisation}
\usage{
run_optimisation(json_args)
}
\arguments{
\item{json_args}{arguments for function \link[kwb.heatsine]{run_optimisation}
in json format}
}
\value{
json list with sim/observation data ("data") fit parameters ("paras"), goodness-of-fit values ("gof")
traveltimes ("traveltimes") and special (min, max, turning) points ("points") as returned by
\link[kwb.heatsine]{get_predictions}
}
\description{
JSON wrapper function for kwb.heatsine::run_optimisation
}
